refactor(#7394): extract Numbers and InlinePhi from Emissions
Move the numeric-literal recipes (number/hex/over-precision checks) and
the inline > [args] only-phi handling out of Emissions into dedicated
Numbers and InlinePhi helpers, clearing qulice's God Class threshold.
